state: let GL scale when displaying savegame thumbnails
No idea why it was done in such a convoluted way. Just removing all the
scaling code, and passing the thumbnail bitmap directly (so GL will
scale it) seems to work.

The only difference I noticed is that it will use bilinear filtering,
while the code before used nearest scaling. This makes it look blurry.
We could easily add a mechanism to prevent the scaling. But it doesn't
bother me, so for now leave it as is.