state: reduce some complicated canvas usage for savcing thumbnails
Until now, this created and entirely new canvas backed by a "software"
bitmap, to render the 3D scene to it. But since we use GL, and there's
no FBO usage or any of that involved, the actual effect is just that we
1. allocate memory for a new bitmap, 2. set the render size/destination.
Both of these are completely unrelated.

Simplify the code to use gr_init_sub_canvas() to set the render targhet
rectangle. (This is similar to how the rear/missile/guidebot windows are
rendered.) Allocate the target bitmap separately.