Remove Screen_mode variable and related things
I think this used to decide what part of the game has focus/should be
rendered. This has been replaced by the window system. The editor also
still sort of uses this, but the editor is broken anyway.

The only place where it still seemed to do anything was init_cockpit().
I replaced the check with whether Game_wind exists, which is probably
correct enough.

The set_screen_mode() calls did effectively nothing, except calling
gr_set_current_canvas(NULL) (this makes the screen canvas active). I
replaced all callers with that, whether or not it actually makes sense.
The resulting behavior should be the same.