Cleanup screen mode business
The Game_screen_mode variable has different meaning from Screen_mode.
The latter is the logical game mode, while the former is the screen
resolution. It's encoded as a single integer using the SM macros. It's
completely unnecessary, because the screen resolution is stored in a
number of other places.

Remove the variables, and cleanup some stuff associated with it. Most
importantly, remove the unnecessary fluff around "render buffers" and
just call init_cockpit() every frame. (I suspect this was not done in
ancient times because they could actually keep the cockpit contents
across multiple frames without redrawing. It doesn't work this way with
GL, so init_cockpit() contained a weird stray draw call.)

Also, the fields that save the screen resolution to the config are now
directly accessed in gr.c. The intention is to save the screen
resolution only if the window is not in fullscreen mode. Though that
doesn't quite work yet (still can save the fullscreen size).