ogl: remove strange buffers
The "pixels" buffer was completely unused.

The "texbuf" buffer was a scratch buffer for texture uploads (since the
texture data needs to be converted or padded anyway). It was allocated
with the size of the screen because this code was shit. Allocate it on
every texture upload instead. It's not going to hurt with the normal
texture sizes the game uses, and at least it's absolutely made sure the
buffer is large enough in every situation.