Remove differences between debug and release mode
In my opinion, release and debug mode shouldn't be too different. Mixing
them just leads to bitrot or unintended behavior.

In fact, a long time ago I was wondering why the hell the main menu
looked different when compiling d2x/dxx in debug mode, and it also
became very crashy. I just want to build the thing in a way I'm able to
use the damn debugger. There are also potentially game play changes due
to corner cases being handled differently. This shouldn't happen.

Remove all RELEASE and NDEBUG ifdefs. Decide on a case by case basis
whether some particular code should remain active or not. Just remove
the code if not. In some cases, actually make this dependent on whether
a runtime "debug mode" flag is on (this is an option in the misc menu,
which is intentionally not saved and restored across program starts).

The only user-visible difference should be the the sandbox main menu
entry, and the debug option in the misc config menu. Enabling the latter
enables weird debug shortcuts, the game time display (why is it actually
there), and some additional runtime checking.

The "heavy-duty error checking" check_segment_connections() makes
loading of Anthology (a very big d2x-xl level) slower by 4.1ms, with -O0
and asan + ubsan enabled.

Note that including additional Asserts or Int3s is not particularly
"dangerous", because these only print messages and then continue the
program. This was changed by me when I was running in debug mode, and
the game was crashing all the time for things that were ignored in
release mode.

link.bat has nothing to do with this, but it needed to die.