ogl: draw rear window borders with lines, instead of pixel loop
This was drawing the thick border around rear/missile/guidebot windows
with a loop that set each pixel separately, with each pixel being a
separate GL_POINTS draw call. What the fuck.

Change to use lines and remove the old code. This looks slightly unclean
if you look at the line joints. This could probably be fixed with some
effort, but in my opinion it's too unnoticeable to waste further time on
it. Besides, stuff like this is probably easier to fix once you have
pixel shaders.