ogl, bitblt: remove duplicated bitmap drawing functions
The only purpose of ogl_ubitblt* is that it doesn't cache the GL texture
in the grs_bitmap struct. It also has a source rect. parameters that are
used by gr_bitmap() for clipping (although effectively, nothing seems to
get clipped at all).

Merge the function with ogl_ubitmapm_cs(). This implicitly means the
texture is now cached in the grs_bitmap passed to it. This could cause
leaks in some cases. movie.c for one creates and destroys a texture on
every frame, so it needs to explicitly free the texture created by the
bitmap drawing function.

movie.c also made it necessary to change the texture filtering handling.
It seems regenerating the mipmap every frame was inefficient enough that
the dxx devs added a user option to disable it (wat). This seems
completely unnecessary. movies only ever get scaled up (due to the low
original resolution of this 1995 game), so mipmaps are useless anyway.
Bilinear scaling on the other hand will always be fast enough (unless
you run _old_ garbage as your GPU). So this commit implicitly drops
this.

I also changed texture filtering to use plain bilinear scaling, instead
of using mapping in any way. Potentially this is visible as worse
downscaling in bilinear mode, but you can just enable the trilinear
scaling option.