render: fix alpha artifacts regressions when weapons hit you
When robots shoot energy weapons at you, a pixel-ish effect gets
rendered over the player's "camera". It looks like a spray of
rectangular solid colored sparks. Let's call them spark bitmaps. They
are rendered with g3_draw_bitmap(), I think. It's important to note that
these bitmaps are rather low resolution (yet rendered over most of the
screen). With texture filtering, there are large translucent regions.

Since commit "New level render order for better seperation and
blending...", multiple overlapping spark textures can show artifacts
typical for reversed order alpha blending, such as being able to see
through the translucent regions, instead of the texture that should be
behind. This can look rather ugly, and even though I noticed that
something was some ugly rendering for brief moments, it took me a while
to reproduce it.

The immediate reason for this regression is that the mentioned commit
removed a GL call to disable writing to the depth buffer from
g3_draw_bitmap(). Readding this (glDepthMask(GL_FALSE) at the start if
the function and undoing it at the end of it) appears to fix it. If it
doesn't write to the depth buffer, other texture fragments rendered in
the wrong order won't be discarded, which makes it look correct wrt.
spark bitmaps. But it could break other weird cases.

It turns out there's a simple fix: sort objects by depth, instead by
distance. This makes it mostly look correct. It's not ideal, because it
could be different from how it's actually rendered (like the weird
offset in draw_object_blob(), rotation, or the GL projection matrix).
Also I'm not sure if the scale of the returned depth is the same (like
the comment on the depth function hints, this does the same as
g3_rotate_point(), except it returns the z coordinate only, so maybe
yes?).

One remaining issue is that if both green and red assholes shoot at you
(D2 level 1), you can see lasers clipped by the spark bitmaps. This
probably requires adjusting sort_func() (which amusingly seems to sort
lasers before spark bitmaps, but whatever).

To test this, it's good if as many robots as possible hit you with
energy weapons. D2 level 1 at highest difficulty level provides plenty
of opportunities.