Remove console and simplify error/warning code
I never used it and it sucks. Remove it. Console messages are now logged
to the terminal. (It doesn't strip certain control characters because
I'm not really seeing the need for it.)

I might readd a better console at some later point if I ever feel like
it.

Reduce the log level of some annoying messages in mveplay.c (why the
fuck did they log these messages with CON_CRITICAL), just because I
noticed these went through the console subsystem.

Also, dxx removed the capability of the game to render warning/error
dialogs with SDL. Instead, they tried to use OS specific code to show a
separate message dialog. (I removed Windows and OSX code for this
earlier; for Linux it was never implemented.) Remove the remaining glue
code for this.

This also means error.c does almost nothing. Remove the messy callback
and global state hell it contained. Just dump errors/warnings to the
terminal.

I want to bring back these in-game rendered message dialogs eventually.
dxx probably removed them because Error() cannot return, which means all
the global state is probably fucked, and it's not a good idea to let the
renderer accidentally call back into the game (or a similar reason).
Warning() will have similar reentrancy issues. My plan is to make all
Error()s non-fatal, and write the messages to a buffer to display them
later when the game (re)enters an event loop. This will require a buffer
mechanism, but that's no reason to keep the current mess.