Bump savegame slots from 10 to 255
There is some brain death involved.

The save/load dialog uses the newmenu.c infrastructure. It creates a
menu entry for each savegame entry. I guess they didn't use the listbox
stuff, because with menu you can "inline" edit entries, which is nice
for savegames.

The first bad thing is that they use a dummy menu entry with a bunch of
newlines and no real text to make space for the thumbnail, and then
render the thumbnail over it. This means the savegame entries start at
index 1, not 0. This breaks scrolling, because the scrolling code
assumes all menu entries have the same height. Move the spacing to the
caption instead. Then scrolling works. Except we need to fix newmenu.c's
determination of the maximum visible entries, which was simply hardcoded
(the added code reduces the count if it overflows the screen size).

To make it worse: it's fairly obvious that the original game devs forgot
about this when they (very) shittily hacked in secret levels. Secret
levels essentially require a separate file for the secret level state,
which is paired with the normal savegame. So for example "pilot.sg0" is
the main savegame, while "1secret.sgc".

Yes, the secret level has an offset of 1, because of the dummy menu
entry above, and the fact that they forgot about it. So this offset has
to be recreated.

You will also notice that the normal savegame has the pilot name in the
filename (in above example just "pilot"), while secret savegame files do
not, so they are shared between multiple pilots. I don't even know why
Descent needs to be able to manage multiple pilot configs.

Even dumber, the last savegame slot uses "asecret.sgc" because for some
reason it uses hex formatting. I'm breaking this with this commit,
because I prefer decimal to keep the filenames slightly saner. If you
try to load an old savegame, and it's in the (formerly) last slot, and
the savegame has a paired secret level, it won't load correctly.

In newgame.c, the added code sets max_on_menu to i+1, which should be an
off-by-1 error. But it looks better as it adds too much padding
otherwise. (I don't know where the padding comes from. Maybe BORDERY is
just that big.) Leaving it as it is, because fuck this shit.

Another thing that could be done is not "returning" the filename by
writing some random string buffer, and using the gen_savegame_fname()
function to recreate it on the caller, but I already felt fucked enough.

255 is an arbitrary number. It allocates a lot of arrays with the
maximum savegame count on the stack, and also loads all thumbnails into
memory, so the number should probably not be too high.