Remove editor files
The editor was part of the original source release, and most likely the
software used for building the official levels. While that sounds
interesting, it's mostly dysfunctional and buggy, and even if it works
it's probably not all that great. Most of the time the editor was
completely broken or didn't even build. (I stopped trying to keep
working what little worked myself.)

DXX upstream has put a lot of effort into restoring its functionality,
but it's still buggy and I didn't manage to use it. (Either because it's
hard to use, or because it crashed before being able to do anything.)
They still call it "!EXPERIMENTAL!". It's not built by default.

What's worse is that the editor is basically part of the game binary,
except if you build as editor, the game may behave subtly different.
While I doubt it changes game play, there is a lot of potential for
unintended behavior. (Might be the reason why DXX builds it as separate
binary in addition to the plain game binary.)

I would say effort should go towards improving DLE[-XP]. Trying to fix
the old editor code is a waste of time. I bet nobody even uses it to
create levels.

Farewell, editor.