piggy: don't crash if there are duplicate read replacement bitmaps
There is at least one D2X-XL level, which has multiple entries that map
to the same bitmap index. This crashed. The previous 2 pass reading code
put an offset into bm->bm_data in the 1st pass, and changed it to a
pointer in the 2nd pass. A duplicate entry would use a pointer as
offset, add it to the base pointer, and write a nonsense data pointer,
which ogl.c would later try to read.

I'm not sure what is the point of the duplicate entries, or if this is
just invalid level data. D2X-XL itself seems to mostly ignore
overwritten bitmaps, but I'm not sure. Change our code to make duplicate
entries overwrite previous entries. Make it 1 pass, and read the bitmap
headers at once before processing the individual bitmaps. (This is
similar to D2X-XL. Alternatively, we could have written the offset to a
separate array, or we could have added seek calls. But this seems
simplest.)