mission, newmenu: improve mission list display
Show the mission type (D1/D2/XL) and the number of levels in a right
column in the list.

Add a way to make the listbox code right-justify half of the entry.

Unfortunately it turns out that the font used is _not_ a monospace font,
so this still looks kind of ugly, unless we add column support to the
listbox renderer. But it's usable.

The mission file parsing code is some serious shitcode. It easily goes
into the list of worst code I've ever had to deal with it. Shit it up
some more because I have only limited time. This fucking garbage should
definitely be rewritten. I can't comprehend how this crap has TWO
duplicated parsers for it, which are both fucking broken and fragile.