dxxerror: add a helper macro
Assert() is currently not fatal (in this code). They often trigger
corner cases that "should not" happen, but which happened anyway, and
just weren't fixed because you usually play the game in release mode,
where Asserts were completely disabled. Without doubt, Assert should be
made fatal again, but this would require removing logging of these
corner cases.

In addition, the original developers _did_ know that their asserts
sometimes failed, and added code to deal with the failure situations in
release builds. Often, the checks were duplicated. (piggy_bitmap_page_in
is a nice example.)

Add a new WARN_IF_FALSE macro that's supposed to help with this. Like
Assert() it dumps useful information about the expression and code
location. It returns the value of the expression itself, so the check
doesn't need to be duplicated.

Also re-document all functions/macros in dxxerror.h.