gamemine: rough hack for d2x-xl reduced shape sides
The D2X-XL mission Kalisto made the game crash (full map cheat + showing
automap was enough). This happened because some segments have invalid
vertex indices. This in turn happened because D2X-XL has a weird, rarely
used feature that allows a level editor to reduce a segment to something
that isn't a cube. Specifically, each side has a "corner" list which
indexes into the segment vertex index list (yeah...), which essentially
means a side can have 3, 2, or even just 1 vertex. Unused vertexes are
set to values larger than Num_vertices, or, depending on weird D2X-XL
details, to unrelated but valid vertexes.

We don't really support this/don't want to support it, because it'd be a
rather intrusive change. It's a small feature that I found only in 1
level so far, in a place where it doesn't seem to be needed. Apply a
gross hack to "fix" the invalid vertex indices. This results in
degenerate side triangles, but still better than crashing.

As implemented, and with the level that was tested, this results in one
segment which looks almost OK (just broken texture coordinates), and one
which looks and behaves really "weird". Just barely OK compared to
crashing or other gross malfunction, and I was getting bored of this, so
trying to fix it ends here. Other levels could have worse outcomes, such
as weirdly unpassable segment connections.

Also, D2X-XL didn't handle this flawlessly. There were bogus triangles
that were rendered even though the entire side was a connection to
another side, and there was broken texture mapping on proper solid sides
too. (Not much better than my hack, even though D2X-XL has native
support for this???)