Merge segment and segment2 structs
Each segment had fields split over struct segment (Segment[]) and struct
segment2 (Segment2s[]). This was a mostly pointless separation. An old
comment (already removed) implied that segment2 was created to make
segment 512 bytes long. Maybe this was for faster access (shift instead
of multiplication to compute array item offset), but this is pure
speculation. It wasn't true anymore either; the segment struct was 528
bytes before this commit.

The separation is awkward, so merge it back. I always thought savegame
stuff prevented this, but apparently this is not the case at all. Most
segment fields are considered static and thus not even saved.