Give each boss its own state in D2X-XL mode
D2X-XL has "full" support for multiple bosses. Each boss has its own
state, in particular each boss has its own teleport segment list. This
is important for game play, because otherwise the boss might teleport
into an unconnected part of the level which you can't even reach.

For classic D2 levels we keep bug compatibility. The importance of this
is unclear, but doing so isn't much effort, so do it anyway.

This also changes the savegame format. We obviously need to be able to
store the per boss state. (To be honest, this could be a bit simpler.
For example, there isn't a real need to store the teleport/gate segment
list, as this could be regenerated on restore. But there are subtle
issues and minor annoyances, so do it anyway.)

As an unrelated change, bump the number of teleportation segments to
1000, because d2x-xl did that. The new Boss_info state is about 200KB.
It's fortunate this isn't 1996, but over 20 years later.

Nothing of this commit is particularly well tested. That comes later.