Very basic and incomplete implementation of D2X-XL object triggers
d2x-xl has an additional set of triggers that are activated by damaging
or destroying robots. In some cases, they're essentially blow up
switches using special models.

The implementation in d2x-xl is all kinds of annoying. You need a
separate trigger array, need to somehow map back the objects to
triggers, and you need to deal with all the weirdness of d2x-xl's
extended trigger system.

This commit doesn't implement most of that. It only does the most simple
change that is sufficient to finish the mission "Anthology" (3rd part).
The only thing implemented is running the trigger when the object is
destroyed. Beyond that, there are weird timers, things that prevent the
trigger from being run, delaying the trigger, reactor triggers, and
more.

Most of the commit is about loading/saving the object trigger data. This
requires a savegame format change (folded into the previous one). I'm
adding a bunch of unused fields that are probably needed for a more
complete implementation of d2x-xl triggers.

I'm thoroughly annoyed that d2x-xl uses a separate array for object
triggers. I'm trying not to do that, and the Triggers array contains the
object triggers as well. But in an attempt to be careful with unwanted
side effects, they essentially form two disjoint regions anyway.