switch: prevent recursion with D2X-XL chained switches
It's possible through TT_MASTER going back to itself. Some d2x-xl levels
do this, and it seems like d2x-xl just added a dumb hack to deal with
this "somehow". Do the same. (At least we're not using a global variable
for this.)

I suppose a proper solution would be either setting a temporary flag in
TT_MASTER switches, or using a bitmap. In theory that could result in
different behavior, but it's unknown whether this would matter for
gameplay anywhere.