collide: fix DXX earthshaker switch blowup regression
DXX upstream commit "Pass correct object to check_effect_blowup" (as
merged in this tree) broke the case when earthshaker missile children
blow up wall switches. Before the aforementioned commit was applied, it
they could not blow wall switches. With the commit applied, they do blow
wall switches, but the switch is not actually triggered. This can lead
to stuck gameplay.

The mentioned commit claims it fixes an original game bug, without
much details. It actually should have been about the exact case of
blowing up switches, but if this case happens, the switch is obviously
not triggered. Not sure why upstream didn't care.

What happens is that this code path simply never calls the trigger code,
even if a switch is blown up. Fix this by having check_effect_blowup()
always take care of triggers. collide_weapon_and_wall(), which calls
this function, must not trigger the switch itself anymore, or it would
trigger it twice.

The code is still a bit messy. For example, check_trigger() has a
duplicated and differently looking check for whether the blower is a
player or guidebot. (Probably could call check_trigger() first and check
its return value, but this would surely cause other problems, so stop
caring for now.)

A third caller of check_effect_blowup() is in the demo code. I have no
idea what it should do or why the heck the demo code should care, but a
quick test seems to indicate it's working.

This bug seems to exist in DXX upstream too (at least in the public
master branch at the time of the commit). The switch straight ahead in
D2 level 1 seems to be a good test case. Note that hitting with the
earthshaker too close to the switch could trigger the normal case,
instead of going through blast_nearby_glass().

The level in which this was important and with which I tested was level
9 in "EQ's sesame set", in the room behind the red door. It's relatively
natural to blast this room with an earthshaker, and it will most time
not hit the large switch in this room directly. It's also not possible
to finish the level in this situation without the bug fix.