Fix and redo rear view "sticky" handling
The rear view key binding has the following semantics: short presses
should toggle the state, long presses should set the state to enabled
while the key is down. This didn't quite work anymore; probably broken
with the recent merges.

Rewrite it and move it to gamecntl.c for unrelated reasons. Note that
DXX upstream also rewrote this code at some point, but this version is
different and tries to do with less state. This is also slightly
abstracted, because I want to reuse it for something else (see next
commit).