gamecntl: make transfer-enegery-to-shield key binding "sticky"
Basically the "T" key now behaves like "R", and toggles the conversion
state on short presses.

Not sure if I'll be happy with this. Maybe I'll add a setting that can
configure whether R and T are sticky, toggle a state, or are active only
while held down. You could also argue that the conversion state should
be either part of the savegame or the user config (or neither).