Replace game exit dialog with a "game menu"
Similar to what upstream DXX did, though I just wrote my own.

The only notable thing is that I'm using a hack to pretend that the menu
gets "replaced" when going to another submenu. Instead of actually
closing the menu, I'm just not drawing it. (E.g. visible when loading a
savegame through the game menu. The confirmation dialog is smaller than
the game menu, so normally you'd see the dialog "stacked" on top of the
game menu.)