Make ammo rack gameplay change optional
Early in this fork, I added a change that extends the upper limits of
the ammo rack. Obviously, this is a game play change. This was fine,
because originally, this was in a dedicated branch for such changes (and
not in master). But now I ended up putting all development into the
current branch, and maintaining a separate branch for such things is a
bother.

Make it user configurable instead. There is now an option that enables
the higher limits, and a number of other things related to ammo rack
handling. With this option off (which is the default), the game is
supposed to behave like the original game.

This gets rid of Primary_ammo_max[] entirely, because it didn't help
much. It was even a bit wrong (gauss has no separate ammo).

Fix some minor issues (that existed with this mode enabled) along the
way.