gamecntl: prevent that some cheats cause "stuck" keys
For example doing "freespace" and then hitting ESC appeared to make the
"e" key binding (banking) to be stuck, until you pressed it again.

This happened because completing the cheat sequence shows a dialog,
which is a new "window". All events (including the key up event for "e")
go to this window. But the window's event loop is run in a blocking
manner and happens in the handler for the "e" down event. Once the
dialog window disappears and control goes back to the game window, the
down event is processed further.

Prevent this by returning true from every cheat, which stops processing
the corresponding key event. Or in other word, the last key of a cheat
is ignored by the game input code. In theory, you would have to do this
only for cheats which show a window, but do it for all for consistency.