ogl: remove broken CULL_FACE state handling from reticle drawing code
This disabled and then enabled GL_CULL_FACE. This makes no sense because
it's a 2D function, and in 2D mode this flag is disabled and must stay
disabled. Just remove it.

This fixes clearing of the _left_ sub-window. This clearing is needed
and observable in the case when any level walls are transparent with no
connected segments.