gamesave: change handling of invalid N_save_pof_names values
Some levels have this. I don't know the exact significance of this.
There is an absurd check, apparently referencing certain level editors.
DXX upstream changed it to something that catches all out of bounds
values, so do something similar. (Didn't research this further and
didn't look for the DXX change for this.)

The affected mission is "Lavafalls Extraction Center".