texmerge: fix hit-testing with rotated top textures
Walls can have an overlaid top texture, which can be rotated in 90°
steps. This needs to be taken care of when doing hit tests. The original
code did hit testing on merged textures, while the newer code tests top
and bottom textures separately. (This was done so textures which don't
need "merging" for rendering don't need to be merged just for hit
testing.)

The new code intended to "undo" the rotation by applying the reverse
transformation on the mapped pixel coordinates in order to map it back
to the top texture. But it turns out this is nonsense.

On merging, a pixel from the top texture is sampled and then "blended"
with the bottom texture pixel. If rotation is involved, the top texture
pixel position is obtained by transforming the bottom texture pixel
position. So this is the same direction of coordinate transformation as
needed for hit testing. In both cases you want to find the top texture
position from the bottom one.

The function check_effect_blowup() basically confirms this and uses the
same logic as the one this commit changes to.

I probably didn't notice this because non-0 orientation is somewhat
rare, and some simple tests appeared to be successful. A place where you
notice it for sure is in "Descent 2: First Strike" level 7, the wall
with overlaid super-transparent grid that covers trigger 14.