texmerge, collide: remove some code duplication
In all texture hit test functions, we need to check the transparency
type at specific texture coordinates. collide.c had this duplicated, and
texmerge_test_pixel() also pretty much duplicated some of it for the
top/bottom cases. Move all this into a tmap_test_pixel() function.

Most or all code always explicitly removed the flags from the top tmap
value only. Or in other words, in theory, the bottom texture could have
a value of higher than 0x3FFF. With the current value of MAX_TEXTURES,
this would lead to out of bounds accesses, so this should not be
possible. Thus the new function handles top and bottom textures the same
way for simplicity.