gamesave: remove bogus unset trigger links
This can happen in "Descent 2: First Strike" level 7, trigger 14. The
wall is indeed set to -1. Remove such bogus wall links.

This fix is of extremely dubious value (and the implementation of it is
ugly and verbose); there are certainly hundreds of other things that
could go much worse. The case in question didn't even trigger a crash
once. (This is with broken level data. Although there are some asserts
which check for 100% broken values, none of them get fixed up, and there
are more cases which are not even checked.)