Fix workarounds in previous commit
The previous commit removes UB in "Descent 2: First Strike" level 7, but
breaks game play in level 11 of the same mission. This time, the force
field guarding the blue key is never removed. Indeed, the linked segment
side has no wall set.

Without the previous commit, it does a bunch of out of bounds accesses,
and then manages to remove the force field anyway. It's pretty clear
from the code that this is an unexpected situation, and it just worked
by coincidence.

Fix this by implementing explicit workarounds for a number of wall
types. Obviously I know only of two places where this bug happened so
far, since I was confused/careless, I attempted to do this for the wrong
trigger types at first. Also, for some triggers it was pretty trivial
anyway, or they already could deal with it. In most cases, bogus
triggers are simply ignored.

The list of triggers which require a wall is complete, as far as
implemented triggers go.

The actual trigger types for this are TT_UNLOCK_DOOR on level 7, and
TT_OPEN_WALL on level 11. Of course the changes are tested only for
these 2 trigger types.

This essentially reverts and replaces the previous commit.