gamecntl: make converter toggle flag part of player state
Turns out having to adjust this flag every time you start the game and
load a savegame is annoying. Also you don't want to have this flag
sticking around when loading another level.

Making this toggle part of the player flags solves this. In particular,
the flag is saved/restored as part of the savegame state, and starting a
new game always resets this.

In general, this behaves quite similar to the headlight toggle. Unlike
the headlight toggle, we don't reset the converter flag if the ship is
destroyed. On one hand, this is intentional (whether you want the toggle
on mostly depends on the kind of level you're playing), on the other
hand, it doesn't matter if the toggle is on even though the player
doesn't have the converter powerup.