state: switching between secret level should keep bomb selection
Going between a level and a secret level is handled via
saving/restoring. This incorrectly included overwriting the current bomb
selection (normal bomb vs. super bomb, controlled via the bomb selection
key).

Don't overwrite this information. To be honest I don't even know why
that Primary/Secondary_last_was_super array even exists an why this
issue only affects the bomb selection, but it should be safe, so
whatever. The logic for (not) restoring this is now the same as with the
afterburn charger state.