ai: explicitly reset awareness stuff on game loading
Not sure if this matters, I added a while ago while trying to debug a
potentially related glitchy thing (robots being aware of you when they
sdhouldn't).

While I'm not sure if this changes anything, it's also not harmful in
any way.

A more correct commit would reset this on normal game start too.