cfile: list opened hogs if all hog losts are used
We have a static limit on the number of hog files opened at the same
time. This is fine, and overflows only happen when there are bugs. This
whole thing is awfully stateful, with no full state wipe when missions
are closed, so there will be bugs. Logging the files helps a little.