mission: reset Robots_replacement_loaded on loading
I have no idea what this variable is for. Hopefully it's some sort of
optimization, so resetting it causes no harm. This was causing problems
because for some absurd reason, load_level_robots() calls
load_mission_ham() again, which causes d2x.hog to be loaded twice in
some situations. Resetting the variable helps avoiding this.

It feels like there's more bugginess behind it (does it really unload
all required state? why is this that stuff in load_level_robots() even
needed?), but for now I'm content.

This happened when starting a d2x-xl level a second time.