cfile: make cfgets() return unterminated strings at the end of the file
If there was a line without a terminating \0 or \n at the end of the
file, this returned NULL instead of the string. Change this and return
the string instead. The returned string is of course correctly \0
terminated.

This is needed for implementing TT_MESSAGE in a convenient way. .msg
files separate records with \n, but the last record may not be
terminated with \n.

Code using this function relies on it returning NULL at the end of the
file, so still return NULL if the returned string is empty. (Users can't
distinguish, nor do they care, whether the last line was terminated or
directly EOF.)