D2X-XL: implement TF_AUTOPLAY, TT_ENABLE_TRIGGER, TT_MESSAGE
This makes it possible to play level 1 of the D2X-XL mission "The
Sphere". (Without the changes part of the gameplay is missing.)

The D2X-XL level format is neither clean, nor is its trigger state
machine simple. Rather, it's very complicated and full of magic numbers
and logic. For this reason, I just straight up ported parts of the
d2x-xl code. Since there is no logic behind this code, it's not really
possible to add support without using its code.

Of course I'm not entirely sure about the correctness. I didn't even
play the level in d2x-xl. But the level appears to make sense now. You
can still leave the level without destroying the reactor, though.