hudmsg: increase message length
One TT_MESSAGE from the previous commit exceeds the default size.
Increase it, so the message is not cut off. You never know with D2, but
in this case no file formats or savegames seem to depend on it, so it
should be harmless.

(I still need to switch to the terminal to actually read the full
message on the terminal. I think d2x-xl implemented TT_MESSAGE as
separate window, not as HUD message. Whatever.)