piggy, texmerge: add a hack for D2X-XL possibly incorrect level flags
The D2X-XL mission "The Sphere", level 1, has an outdoor room, from
which you're supposed to see the skybox. We rendered a white opaque
texture instead. I'm not fully sure why this happens, while d2x-xl is
supposedly fine (couldn't test it).

It turns out it uses a replacement texture, which uses the paletted
format, and where every pixel is 255, but BM_FLAG_TRANSPARENT is not
set. What the heck. I couldn't find out whether d2x-xl actually sets it.
It has plenty of code to compute the flags for RGBA textures, but
paletted textures are handled differently. It might be that the code
accidentally sets all transparency flags:

  pBm->SetTranspType ((pBm->Flags () | (BM_FLAG_TRANSPARENT | BM_FLAG_SUPER_TRANSPARENT)) ? 3 : 0);

Is there a '&' vs. '|' typo? But not sure if this is really on the
relevant code path. If it is, d2x-xl just considers all textures
transparent and super-transparent.

Anyway, add a hack to set the flags from the texture data. Specifically
run this only with d2x-xl (and only from .pog files), since it has
potential for regressions (palette entry 254 can be used either for a
real color, or for super transparency, depending on the bitmap flags),
and it's probably slower. Reuse the code for RGBA textures, it's not
much top make it generic (since actual tex merging and hit testing also
use parts of this).