render: fix missing view matrix transformation if render depth reached
If the Render_depth is reached, and there were still new segments to
process, they didn't get rotated correctly, and non-sense was rendered.
This is especially visible when setting Render_depth to 1 or 2.

Fix this by explicitly transforming unprocessed points at the end. (You
could also do this when adding new segments to the render list, but I
guess the current way reduces the number of rotate_list() call sites
most. The sky box one can be folded into the new call too.)