collide: respect TF_PERMANENT
D2X-XL has "permanent" switches, which can be retriggered any time.
Apparently this will just keep the switch texture in place (I haven't
tested it with d2x-xl, but it seems to work this way).

Unfortunately, this sucks quite a bit. If you fire with a quad laser at
such a switch, it can trigger up to 4 times. If this were properly done,
a switch would have a transition animation between two states, and
couldn't be retriggered during that animation. Oh well, even if it's
such an awful hack, some D2X-XL levels need it.

There's unfortunately a bit more to to TF_PERMANENT (did I mentioned yet
what an awful fuckmess D2X-XL does out of switches?), which is untested
and was added in earlier commit, but I guess that provides fun for
future debugging.